pyproject.toml
[project]
name = "desktop-searcher"
version = "0.1.0"
description = "Starter Python project managed by uv"
requires-python = ">=3.9"
readme = "README.md"
authors = [{ name = "Your Name" }]
license = { text = "UNLICENSED" }
dependencies = [
    "duckdb>=1.4.3",
    "httpx>=0.28.1",
    "lightgbm>=4.6.0",
    "loguru>=0.7.3",
    "matplotlib>=3.9.4",
    "numpy>=2.0.2",
    "pandas>=2.3.3",
    "polars>=1.36.1",
    "psycopg2-binary>=2.9.11",
    "pyarrow>=21.0.0",
    "pydantic>=2.12.5",
    "pypdf>=6.5.0",
    "python-docx>=1.2.0",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "rich>=14.2.0",
    "rumps>=0.4.0",
    "scikit-learn>=1.6.1",
    "scipy>=1.13.1",
    "seaborn>=0.13.2",
    "sentence-transformers>=5.1.2",
    "sqlalchemy>=2.0.45",
    "statsmodels>=0.14.6",
    "torch>=2.1",
    "tqdm>=4.67.1",
    "typer>=0.20.1",
    "watchdog>=6.0.0",
    "xgboost>=2.1.4",
    # Analytics dashboard
    "streamlit>=1.50.0",
    "plotly>=6.5.0",
    # macOS UI + WebKit (only on macOS)
    "pyobjc-core>=11.1; sys_platform == 'darwin'",
    "pyobjc-framework-cocoa>=11.1; sys_platform == 'darwin'",
    "pyobjc-framework-webkit>=11.1; sys_platform == 'darwin'",
    # OCR backends (optional, works cross-platform)
    "pytesseract>=0.3.10",
    "Pillow>=11.0.0",
    # macOS Vision OCR (only on macOS)
    "pyobjc-framework-Vision>=11.1; sys_platform == 'darwin'",
]

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[tool.uv]

[dependency-groups]
dev = [
    "black>=25.11.0",
    "ipykernel>=6.31.0",
    "jupyterlab>=4.5.1",
    "mypy>=1.19.1",
    "pre-commit>=4.3.0",
    "pytest>=8.4.2",
    "ruff>=0.14.10",
]
# uv will use this venv automatically when present
